requirements.txt
# Core dependencies for Diffusion Thought Tensor Research

# Deep Learning Frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# NLP and Diffusion Libraries
transformers==4.36.0
diffusers==0.24.0
datasets==2.16.0
tokenizers>=0.15.0

# Training and Optimization
accelerate==0.25.0
pytorch-lightning==2.1.3
wandb==0.16.1
tensorboard==2.15.0

# Advanced Transformers
x-transformers==1.27.0
einops==0.7.0

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# Scientific Computing
numpy>=1.24.0
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly>=5.18.0

# Utilities
tqdm==4.66.1
pyyaml>=6.0
pandas>=2.1.0

# Development Tools
jupyter==1.0.0
ipykernel==6.28.0
black==23.12.1
flake8==7.0.0
pytest==7.4.3
ipywidgets>=8.1.0

# Optional: For MDLM reference implementation
# git+https://github.com/kuleshov-group/mdlm.git